Anjay 2.3a
BREAKING CHANGES:
- Removed usages of the ssize_t type. APIs in both Anjay and avs_commons that
  had it in public signatures have been redesigned
- (commercial version only) Retry mechanisms described in LwM2M TS 1.1.1,
  section 6.2.1.2 are now used by default, which changes the default
  registration retry policy
- Updated avs_commons to version 4.1, with the following breaking changes:
  - Renamed public header files for better uniqueness
  - Redesigned socket creation and in-place decoration APIs, including the
    addition of a requirement to provide PRNG context
  - Renamed some public configuration macros, to unify with the updated
    compile-time configuration pattern
  - Removed the legacy avs_coap component (the version used by Anjay 1.x)
  - Removed the mbed TLS custom entropy initializer pattern in favor of the new
    PRNG framework

Note: For a more detailed information about breaking changes and how your code
needs to be fixed, see:
https://avsystem.github.io/Anjay-doc/Migrating/MigratingFromAnjay225.html

Features:
- Changed project structure, configuration headers and updated build system, so
  that building the library without using CMake is now officially supported
- Allowed public access to Anjay's scheduler using anjay_get_scheduler()
- Code generator now allows omitting some of the resources during generation
- (commercial version only) Added support for some of the retry mechanisms
  described in LwM2M TS 1.1.1, including the following resources:
  - Bootstrap on Registration Failure
  - Communication Retry Count
  - Communication Retry Timer
  - Communication Sequence Delay Timer
  - Communication Sequence Retry Count
- (commercial version only) SMS driver API is now public, allowing for custom
  driver implementation

Improvements:
- Major overhaul of documentation, including new tutorials for data model and
  firmware update implementation, as well as guides for migration from both
  Anjay 1.16 and 2.2
- Cryptographically secure PRNG is now used whenever possible for generation of
  CoAP tokens and initial message IDs
- Improvements to code generator:
  - Refactored the generated code so that programmatic instantiation of
    Object Instances is now easier
  - Code generated in C++ mode is now more object-oriented and idiomatic
- Thanks to all compile-time configuration now being accessible via a public
  header, the demo client can now be compiled even when some optional features
  are disabled
- Removed some unused code from the open source version

Bugfixes:
- Fixed a problem that could occur while reconnecting to a server, if the host's
  local address was switching between IPv4 and IPv6
- Fixed behavior of anjay_exit_offline() when called immediately after
  anjay_enter_offline()
- disable_legacy_server_initiated_bootstrap is now properly respected when
  reconnecting after a connectivity failure
- anjay_get_string() and anjay_execute_get_arg_value() now report buffer
  underruns more reliably
- Added missing HTTP download timeout logic
- Errors from the socket's send() method are now properly propagated through the
  CoAP layer
- Fixed fatal errors that could occur during some specific CoAP error
  conditions, including during attempts to silently ignore incoming packets
- Fixes to minor issues found by Coverity
- Fixed working on platforms where malloc/calloc returns NULL when 0 bytes is
  requested
- Fixed compatibility with some embedded compilers
- Fixed interoperability with servers that use Uri-Path: '' to represent empty
  query path, as it seems to be permitted by the CoAP RFC
- Fixed error codes used by the Portfolio object in the demo client
- Minor fixes in integration testing framework
- (commercial version only) Added missing support of LwM2M 1.1-specific
  resources of the Server object to its persistence functions
- (commercial version only) Fixed various bugs in the bootstrapper module
- (commercial version only) Additional verification of NIDD URLs
- (commercial version only) More robust error handling in demo client's NIDD
  driver

Also updates avs_commons to version 4.1, which includes the following changes,
in addition to those mentioned above as "breaking changes":

Features:
- Building without CMake is now officially supported
- Added idiomatic C++ wrapper for AVS_LIST
- New API for cryptographically safe PRNGs in avs_crypto
- File-based streams and default log handler can now be disabled at compile time

Bugfixes:
- Fixed a bug in the default socket implementation that prevented compiling on
  platforms without the IP_TOS socket option support
- Fixed improper parsing of empty host in URLs
- Some previously missed log messages now properly respect WITH_AVS_MICRO_LOGS
- Fixed a bug in netbuf stream's error handling

+++ demo/objects/apn_conn_profile.c
@@ -207,9 +207,9 @@ static int apncp_resource_write(anjay_t *anjay,
             return ANJAY_ERR_INTERNAL;
         }
 
-        ssize_t result = snprintf(inst->profile_name,
-                                  sizeof(inst->profile_name), "%s", buf);
-        if (result < 0 || result >= (ssize_t) sizeof(inst->profile_name)) {
+        int result = snprintf(inst->profile_name, sizeof(inst->profile_name),
+                              "%s", buf);
+        if (result < 0 || (size_t) result >= sizeof(inst->profile_name)) {
             return ANJAY_ERR_INTERNAL;
         }
         inst->has_profile_name = true;
